pyproject.toml
[tool.poetry]
name = "lola"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.94.1"
httpx = "^0.23.3"
uvicorn = "^0.21.1"
google-api-python-client = "^2.81.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^1.0.0"
pandas = "^1.5.3"
requests = "^2.28.2"
dockerignore-generate = "^1.0.7"
pydantic = "^1.10.7"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.2.0"
pytest = "^7.2.2"
black = "^23.1.0"
flake8 = "^6.0.0"
mypy = "^1.1.1"
isort = "^5.12.0"
pycodestyle = "^2.10.0"
pactman = "^2.30.0"
vcrpy = "^4.2.1"


[tool.isort]
force_single_line = true
line_length = 120
skip_glob = "**/migrations/*.py"
lines_between_types = 1
known_fastapi = "fastapi"
known_third_party = ["urllib3", "pytest"]
sections = ["FUTURE", "STDLIB", "FASTAPI", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]


[tool.black]
line-length = 120
target_version = ['py39']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | migrations
  )/
)
'''


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
